checking in remove herdocs for NULL. Providing no such file or directory error message if path is unset

+++ sources/main.c
@@ -6,7 +6,7 @@
 /*   By: cmenke <cmenke@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/25 13:48:03 by cmenke            #+#    #+#             */
-/*   Updated: 2023/07/31 19:43:20 by cmenke           ###   ########.fr       */
+/*   Updated: 2023/08/01 11:41:25 by cmenke           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -186,6 +186,8 @@ void	ft_delete_heredoc_files(char **heredoc_files)
 	int	i;
 
 	i = 0;
+	if (!heredoc_files)
+		return ;
 	while (heredoc_files[i])
 	{
 		if (unlink(heredoc_files[i]) == -1)
